alphalens-mvp/requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.2.0
websocket-client>=1.7.0
openai>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
